op_alloy::consensus
Enum OpReceiptEnvelope

#[non_exhaustive]
pub enum OpReceiptEnvelope<T = Log> {
    Legacy(ReceiptWithBloom<T>),
    Eip2930(ReceiptWithBloom<T>),
    Eip1559(ReceiptWithBloom<T>),
    Eip7702(ReceiptWithBloom<T>),
    Deposit(OpDepositReceiptWithBloom<T>),
}

Available on crate feature consensus only.

Receipt envelope, as defined in EIP-2718, modified for OP Stack chains.

This enum distinguishes between tagged and untagged legacy receipts, as the in-protocol merkle tree may commit to EITHER 0-prefixed or raw. Therefore we must ensure that encoding returns the precise byte-array that was decoded, preserving the presence or absence of the TransactionType flag.

Transaction receipt payloads are specified in their respective EIPs.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Legacy(ReceiptWithBloom<T>)

Receipt envelope with no type flag.

Eip2930(ReceiptWithBloom<T>)

Receipt envelope with type flag 1, containing a EIP-2930 receipt.

Eip1559(ReceiptWithBloom<T>)

Receipt envelope with type flag 2, containing a EIP-1559 receipt.

Eip7702(ReceiptWithBloom<T>)

Receipt envelope with type flag 4, containing a EIP-7702 receipt.

Deposit(OpDepositReceiptWithBloom<T>)

Receipt envelope with type flag 126, containing a deposit receipt.

Implementations

impl OpReceiptEnvelope

pub fn from_parts<'a>( status: bool, cumulative_gas_used: u128, logs: impl IntoIterator<Item = &'a Log>, tx_type: OpTxType, deposit_nonce: Option<u64>, deposit_receipt_version: Option<u64>, ) -> OpReceiptEnvelope

Creates a new OpReceiptEnvelope from the given parts.

impl<T> OpReceiptEnvelope<T>

pub const fn tx_type(&self) -> OpTxType

Return the OpTxType of the inner receipt.

pub fn is_success(&self) -> bool

Return true if the transaction was successful.

pub fn status(&self) -> bool

Returns the success status of the receipt’s transaction.

pub fn cumulative_gas_used(&self) -> u128

Returns the cumulative gas used at this receipt.

pub fn logs(&self) -> &[T]

Return the receipt logs.

pub const fn logs_bloom(&self) -> &Bloom

Return the receipt’s bloom.

pub fn deposit_nonce(&self) -> Option<u64>

Return the receipt’s deposit_nonce if it is a deposit receipt.

pub fn deposit_receipt_version(&self) -> Option<u64>

Return the receipt’s deposit version if it is a deposit receipt.

pub const fn as_deposit_receipt_with_bloom( &self, ) -> Option<&OpDepositReceiptWithBloom<T>>

Returns the deposit receipt if it is a deposit receipt.

pub const fn as_deposit_receipt(&self) -> Option<&OpDepositReceipt<T>>

Returns the deposit receipt if it is a deposit receipt.

pub const fn as_receipt(&self) -> Option<&Receipt<T>>

Return the inner receipt. Currently this is infallible, however, future receipt types may be added.

impl OpReceiptEnvelope

pub fn inner_length(&self) -> usize

Get the length of the inner receipt in the 2718 encoding.

pub fn rlp_payload_length(&self) -> usize

Calculate the length of the rlp payload of the network encoded receipt.