op_alloy::consensus::hardforks
Module ecotone

Available on crate feature consensus only.

Module containing a Transaction builder for the Ecotone network updgrade transactions.

Constants

EIP4788_FROM

EIP-4788 From Address

ENABLE_ECOTONE_INPUT

The Enable Ecotone Input Method 4Byte Signature

GAS_PRICE_ORACLE_ADDRESS

The Gas Price Oracle Address This is computed by using go-ethereum’s crypto.CreateAddress function, with the Gas Price Oracle Deployer Address and nonce 0.

GAS_PRICE_ORACLE_DEPLOYER_ADDRESS

The Gas Price Oracle Deployer Address

L1_BLOCK_DEPLOYER_ADDRESS

L1 Block Deployer Address

NEW_L1_BLOCK_ADDRESS

The new L1 Block Address This is computed by using go-ethereum’s crypto.CreateAddress function, with the L1 Block Deployer Address and nonce 0.

UPDGRADE_TO_FUNC_SIGNATURE

The UpdgradeTo Function Signature