op_alloy::consensus::hardforks
Module fjord

Available on crate feature consensus only.

Module containing a Transaction builder for the Fjord network upgrade transactions.

Constants

ENABLE_FJORD_FUNC_SIGNATURE

Input data for setting the Fjord Gas Price Oracle.

FJORD_GAS_PRICE_ORACLE_ADDRESS

Fjord Gas Price Oracle address.

GAS_PRICE_ORACLE_ADDRESS

The Gas Price Oracle Address This is computed by using go-ethereum’s crypto.CreateAddress function, with the Gas Price Oracle Deployer Address and nonce 0.

GAS_PRICE_ORACLE_FJORD_DEPLOYER_ADDRESS

Fjord Gas Price Oracle Deployer Address.

L1_INFO_DEPOSITER_ADDRESS

The L1 Info Depositer Address.

SET_FJORD_METHOD_SIGNATURE

The Set Fjord Four Byte Method Signature.