op_alloy::consensus
Struct TxDeposit

pub struct TxDeposit {
    pub source_hash: FixedBytes<32>,
    pub from: Address,
    pub to: TxKind,
    pub mint: Option<u128>,
    pub value: Uint<256, 4>,
    pub gas_limit: u64,
    pub is_system_transaction: bool,
    pub input: Bytes,
}

Available on crate feature consensus only.

Deposit transactions, also known as deposits are initiated on L1, and executed on L2.

Fields
source_hash: FixedBytes<32>

Hash that uniquely identifies the source of the deposit.

from: Address

The address of the sender account.

to: TxKind

The address of the recipient account, or the null (zero-length) address if the deposited transaction is a contract creation.

mint: Option<u128>

The ETH value to mint on L2.

value: Uint<256, 4>

The ETH value to send to the recipient account.

gas_limit: u64

The gas limit for the L2 transaction.

is_system_transaction: bool

Field indicating if this transaction is exempt from the L2 gas limit.

input: Bytes

Input has two uses depending if transaction is Create or Call (if to field is None or Some).

Implementations

impl TxDeposit

pub fn decode_fields(buf: &mut &[u8]) -> Result<TxDeposit, Error>

Decodes the inner TxDeposit fields from RLP bytes.

NOTE: This assumes a RLP header has already been decoded, and just decodes the following RLP fields in the following order:

source_hash

from

to

mint

value

gas_limit

is_system_transaction

input

pub fn size(&self) -> usize

Calculates a heuristic for the in-memory size of the TxDeposit transaction.

pub fn encode_inner(&self, out: &mut dyn BufMut, with_header: bool)

Inner encoding function that is used for both rlp [Encodable] trait and for calculating hash that for eip2718 does not require rlp header

pub fn encoded_len(&self, with_header: bool) -> usize

Output the length of the RLP signed transaction encoding.

If with_header is true, the length includes the RLP header.

pub fn signature() -> Signature

Returns the signature for the optimism deposit transactions, which don’t include a signature.