op_alloy::protocol
Trait BatchValidationProvider

pub trait BatchValidationProvider {
    type Error: Display + ToString;

    // Required methods
    fn l2_block_info_by_number<'life0, 'async_trait>(
        &'life0 mut self,
        number: u64,
    ) -> Pin<Box<dyn Future<Output = Result<L2BlockInfo, Self::Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn block_by_number<'life0, 'async_trait>(
        &'life0 mut self,
        number: u64,
    ) -> Pin<Box<dyn Future<Output = Result<Block<OpTxEnvelope>, Self::Error>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
}

Available on crate feature protocol only.

Describes the functionality of a data source that fetches safe blocks.

Required Associated Types

type Error: Display + ToString

The error type for the BatchValidationProvider.

Required Methods

fn l2_block_info_by_number<'life0, 'async_trait>( &'life0 mut self, number: u64, ) -> Pin<Box<dyn Future<Output = Result<L2BlockInfo, Self::Error>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Returns the L2BlockInfo given a block number.

Errors if the block does not exist.

fn block_by_number<'life0, 'async_trait>( &'life0 mut self, number: u64, ) -> Pin<Box<dyn Future<Output = Result<Block<OpTxEnvelope>, Self::Error>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Returns the OpBlock for a given number.

Errors if no block is available for the given block number.