op_alloy::rpc_types
Module net

Available on crate feature rpc-types only.

Network RPC types

Structs

GossipScores

PeerDump

PeerInfo

PeerScores

PeerStats

ReqRespScores

TopicScores

Enums

Connectedness

Represents the connectivity state of a peer in a network, indicating the reachability and interaction status of a node with its peers.

Direction

Direction represents the direction of a connection.